The refresh rate in SDL_DisplayMode is now a float

+++ include/SDL3/SDL_test_common.h
@@ -78,7 +78,7 @@ typedef struct
     int logical_h;
     float scale;
     int depth;
-    int refresh_rate;
+    float refresh_rate;
     int num_windows;
     SDL_Window **windows;
